Fix tagging in maven workflow
Using the docker/metadata-action generates tags with full image name.
This makes it incompatible with jibs job.to.tags parameter which require
only the tag part. To make things simple the version calculated by
the metadata action is used as tag and by default jib pushes a latest tag
when the image is specified without a tag.